Grammar changes and addition of new paragraph
This is the first draft of adding 'HTML Injection' context into the Angular security documentation. The documentation covers a wide variety of XSS prevention but does not talk about using the correct template context. If `innerHTML` is used where maybe an expression or `innerText` should have been used, whilst an attacker cannot inject JavaScript, it may be possible to inject valid HTML in attempts to get a user to click a link to a third-party website. 

Writing this because this is the common issue I see with angular applications for many clients, a few examples is when they are displaying changes, or showing a name; which can be controlled by an attacker allowing for HTML Injection and can be seen by more than one person in the application.